VIM: Better reply handling with multiple emails
This patch fixes reply handling when using multiple emails.  This
adds a config check for other_email and uses that information when
formulating reply headers.  It will strip out your own email addresses
from the reply and set the From: to be an email of yours found in the
original message.